Remove storybook
We've tried during more than a week, to migrate storybook from v5 to v8
in order to fix Node 20 support (after upgrading webpack to v5), but
failed.
There are too many breaking changes everywhere and this project is full
of tech debt, see storybook4-state dependency.

It costs more than it brings in.

So, we've decided to remove the storybook, before upgrading to webpack
5, which will fix the Node 20 support.
And we'll bring back the "demos" later, in a KISS way.